Add override to extract_library endpoint.

+++ app/api/routes/code.py
@@ -10,6 +10,7 @@ import logging
 import shutil
 
 from app.databases.singletons import get_mongo_db
+from app.models.code import CodeContent, CodeChunk, FinalCodeChunk, GitUrl
 
 logging.basicConfig(level=logging.DEBUG)
 
@@ -18,13 +19,24 @@ router = APIRouter()
 mdb_dep = Annotated[MongoDBDatabase, Depends(get_mongo_db)]
 
 @router.get("/extract_library/")
-async def clone_library(git_url: str,override: bool, mdb: mdb_dep):
+async def extract_library(git_url: str,override: bool, mdb: mdb_dep):
     try:
-        folder_path = await clone_git_repo(mdb,git_url,override)
-        print(folder_path)
-        await extract_contents(folder_path, git_url)
-        shutil.rmtree(folder_path)
-        return {"status": "success", "message": "Library cloned and processed successfully."}
+        if override:
+            await mdb.delete_entries(CodeContent, doc_filter={"url": git_url})
+            await mdb.delete_entries(CodeChunk, doc_filter={"url": git_url})
+            await mdb.delete_entries(FinalCodeChunk, doc_filter={"url": git_url})
+            await mdb.delete_entries(GitUrl, doc_filter={"url": git_url})
+
+        urls = await mdb.get_entries(GitUrl, doc_filter={"url": git_url})
+        if len(urls) == 0:
+            await mdb.add_entry(GitUrl(url=git_url))
+            folder_path = await clone_git_repo(mdb,git_url,override)
+            print(folder_path)
+            await extract_contents(folder_path, git_url)
+            shutil.rmtree(folder_path)
+            return {"status": "success", "message": "Library cloned and processed successfully."}
+        else:
+            return {"status": "success", "message": "Library is already cloned"}
     except Exception as e:
         logging.exception("Error cloning library")
         raise HTTPException(status_code=500, detail=str(e))
